+++ test/matchers.spec.ts
@@ -1,6 +1,9 @@
 import {DetailedCellError} from '../src'
 import {CellError, ErrorType} from '../src/Cell'
 import {adr} from './testUtils'
+import {FormulaVertex} from '../src/DependencyGraph/FormulaCellVertex'
+import {buildNumberAst} from '../src/parser/Ast'
+import {MatrixSize} from '../src/MatrixSize'
 
 describe('Matchers', () => {
   it('should compare two simple values', () => {
@@ -9,23 +12,27 @@ describe('Matchers', () => {
   })
 
 
-  it('should compare two cell errors ignoring addresses', () => {
+  it('should compare two cell errors ignoring vertices', () => {
+    function dummyFormulaVertex(): FormulaVertex {
+      return FormulaVertex.fromAst(buildNumberAst(1), adr('A1'), MatrixSize.scalar(), 0)
+    }
+
     expect(
-      new CellError(ErrorType.ERROR, '', adr('A1'))
+      new CellError(ErrorType.ERROR, '', dummyFormulaVertex())
     ).toEqualError(
       new CellError(ErrorType.ERROR, '')
     )
 
     expect(
-      new CellError(ErrorType.ERROR, 'a', adr('A1'))
+      new CellError(ErrorType.ERROR, 'a', dummyFormulaVertex())
     ).not.toEqualError(
-      new CellError(ErrorType.ERROR, '', adr('A1'))
+      new CellError(ErrorType.ERROR, '', dummyFormulaVertex())
     )
 
     expect(
-      new CellError(ErrorType.NA, '', adr('A1'))
+      new CellError(ErrorType.NA, '', dummyFormulaVertex())
     ).not.toEqualError(
-      new CellError(ErrorType.ERROR, '', adr('A1'))
+      new CellError(ErrorType.ERROR, '', dummyFormulaVertex())
     )
   })